Move findPrNumberAndHostingService to GitHistoryService
