+++ src/main/kotlin/com/github/shiraji/findpullrequest/model/GitHistoryService.kt
@@ -1,5 +1,6 @@
 package com.github.shiraji.findpullrequest.model
 
+import com.github.shiraji.getNumberFromCommitMessage
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vcs.history.VcsRevisionNumber
 import git4idea.GitCommit
@@ -25,5 +26,14 @@ class GitHistoryService {
 
     fun hasCommitsFromRevisionNumber(commits: List<GitCommit>, revisionHash: VcsRevisionNumber): Boolean {
         return commits.any { it.id.asString() == revisionHash.asString() }
+    }
+
+    fun findPrNumberAndHostingService(pullRequestCommit: GitCommit): Pair<Int?, FindPullRequestHostingServices?> {
+        var prNumber: Int? = null
+        val targetHostingService = FindPullRequestHostingServices.values().firstOrNull {
+            prNumber = pullRequestCommit.getNumberFromCommitMessage(it.defaultMergeCommitMessage)
+            prNumber != null
+        }
+        return Pair(prNumber, targetHostingService)
     }
 }
